Fix evaluating inside IIFEs
Problem was twofold: first could have multiple markers
associated with the same evaluation comment.  But only one them would
receive a result back from the worker.  So `undefined`.

I fixed this by using a map to keep track of comments already
seen and the marker we constructed for it.

The we did not associate evaluation comments to the nearest parent
expression statement, but to all expression statements above.  Thus, in
an IIFE like:

(function(x) {
  x //:
})(1)

there would be two ExpressionStatements: `x`, and the IIFE.  Both would
be associated to the one evaluation comment, and receive a result from
the worker.  And the second result would overwrite the first, so
`undefined`.

I fixed that by doing a first traversal of the AST to find evaluation
comments and associate them with the nearest parent expression
statement.