Fix the //: delimiter showing all values of the current iteration
On the first run.  This was due to having no characters after the delimiter,
which meant the `from` and `to` markers were identical, and this ended up
inserting the text and never erasing the rest of the line.

The fix is to use `null` as the character value, which CodeMirror takes to mean
"until the end of the line".